Starting work on networking.
Networking is hard. Best to take care of it early.
Otherwise I may end up in the position of many other games where
'It would be great to have a multiplayer version of the game,
but it is just too much work trying to get it integrated into the
current game engine.'